Fix errors in query recordings mounted handler
The guard for the query in the `mounted` handler was not complete
meaning there were errors reported every time the page loaded.

The empty object is now checked for correctly and there specific
guards for each query field in case any field is missing.

Also changed behaviour so that a query is always automatically
performed on page load as this has been requested before.