Restore recording comment functionality
A new Comment button has been added which pops up a modal which shows
the current comment and allows it to be updated. Comment updates are
pushed through the store (the previous implementation made a direct
API request).

Four buttons on a single row didn't render well on mobile so the
buttons now spread over two rows on smaller screens.