Fix bug where fields to update were being wrapped in a fields object
Intended for CacophonyApi.patch to receive the exact object as passed into updateFields,
for example if updateFields given:
{ email: "mail@jacksteel.co.nz" }
It should pass that to PATCH body.
Instead, it was incorrectly wrapping it into a fields object, meaning the PATCH body looked like
{ fields : { email: "mail@jacksteel.co.nz" } }
Which isn't what it should be and isn't what the server expects, meaning PATCH requests were failing.